+++ entries/common.entry.js
@@ -2,6 +2,7 @@ import ajaxIconSprite from '@larva-js/utils/ajaxIconSprite';
 import initFlickity from '@larva-js/interface/Flickity';
 import initVideoShowcase from '@larva-js/interface/VideoShowcase';
 import initCollapsible from '@larva-js/interface/Collapsible';
+import initPopOver from '@larva-js/interface/PopOver';
 
 const spritePath = ( () => {
 	if ( undefined !== window.pmc_common_urls ) {
@@ -20,4 +21,5 @@ document.addEventListener( 'DOMContentLoaded', () => {
 	initCollapsible();
 	initFlickity();
 	initVideoShowcase();
+	initPopOver();
 } );