+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished
+* larva-patterns - Add `author` and `author-details` module.
+* larva-tokens - Update grey color for rollingstone-2022.
 * larva-patterns - Add `article-title` module.
 * larva-patterns - Add `article-kicker` module.
 * larva-patterns - Add `article-excerpt` module.