+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished
+* larva-patterns - `story-grid` module changes. 
+* larva-css - Add a-list algorithm.
+* larva-css - Add new font size.
+
+## Unpublished
 * larva-patterns - Add `contentt-showcase` module.
 * larva-patterns - Add `content-showcase-grid` module.
 * larva-css - Add a-flex algorithm.