+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished
-* larva-css - Add `lrv-u-width-35\@mobile-max`.
+* larva-css - Add `lrv-u-width-50` & `lrv-u-width-35\@mobile-max`.
 
 ## Unpublished
 * larva-patterns - Add `author` and `author-details` module.