+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unpublished
+## 0.5.6 - 03-29-2022
 * larva-tokens - Add desktop_xl token values in typography.json
 * larva-tokens - Add desktop_xl breakpoint in font-data.js
 * larva-css - Update css generator script to include desktop_xl breakpoint using media queries